scheduler/docket/requirements.txt
amqp==2.2.2
asn1crypto==0.24.0
billiard==3.5.0.3
celery==4.1.0
cffi==1.11.4
cryptography==2.1.4
Django==2.0.2
django-celery-beat==1.1.1
djangorestframework==3.7.3
ephem==3.7.6.0
idna==2.6
kombu==4.1.0
psycopg2==2.7.3.2
pycparser==2.18
pymongo==3.6.0
PyMySQL==0.8.0
pytz==2017.3
six==1.11.0
SQLAlchemy==1.2.2
vine==1.1.4
uWSGI==2.0.15
xlrd==1.1.0
boto3==1.6.12